train_test_split
test_size = 0.2 means 20% Testing Data and 80% Training Data

stratify = we need almost equal number of rocks and mines. Data will be splitted based on the number of rocks and mines

random_state = It is the order in which the data is in a particular order, the number will decide in what way the training data example will be selected.